requirements.txt
configparser==3.7.4
Django==2.2.2
django-bootstrap-modal-forms==1.4.2
django-crispy-forms==1.7.2
django-simple-history==2.7.3
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.1
six==1.12.0
South==1.0.2
sqlparse==0.3.0
